Finished off differentiating code/list blocks
Requires the actual saving of the file now, also can go back and edit
the title and stuff. Still a little rough around the edges on that...
is there a way to abstract that/not make it all based on globals? If it
was de-globalfied (which is probably really easy I'm just being lazy...
make it into a class???) it could be very easily abstracted. Maybe I
should make all of the fields in this thing into a class huh. Maybe.
Just maybe.